+++ src/codegen_support/runtime/emitters.rs
@@ -1,6 +1,6 @@
 //! Purpose:
 //! Coordinates emission of all runtime helper labels for supported targets.
-//! Orders strings, system helpers, exceptions, arrays, buffers, I/O, pointers, and fibers so dependencies are available.
+//! Orders core, managed-value, and platform-facing helper groups so dependencies are available.
 //!
 //! Called from:
 //! - `crate::codegen_support::runtime::emit_runtime()`.
@@ -8,31 +8,16 @@
 //! Key details:
 //! - Emission order is part of the runtime contract because helpers branch to labels and data symbols emitted elsewhere.
 
-use super::arrays;
-use super::buffers;
-use super::callables;
-use super::diagnostics;
-use super::eval_bridge;
-use super::eval_scope;
-use super::exceptions;
-use super::fibers;
-use super::generators;
-use super::io;
-use super::objects;
-use super::pointers;
-use super::resource_ids;
-use super::spl;
-use super::strings;
-use super::system;
-use super::zval;
+mod managed;
+mod platform;
+
+use super::{callables, diagnostics, exceptions, generators, strings, system};
 use crate::codegen_support::emit::Emitter;
 use crate::codegen_support::RuntimeFeatures;
 
 /// Emits all runtime helper labels in dependency order for supported targets.
 ///
-/// Emits in order: diagnostics, strings, callables, system, exceptions, generators,
-/// arrays, SPL, objects, buffers, I/O, pointers, fibers.
-///
+/// Emits core helpers first, followed by managed-value helpers and platform-facing helpers.
 /// Each category is emitted before any code that depends on it, ensuring labels
 /// are available when branches are assembled.
 pub(crate) fn emit_runtime(emitter: &mut Emitter, features: RuntimeFeatures) {
@@ -184,662 +169,9 @@ pub(crate) fn emit_runtime(emitter: &mut Emitter, features: RuntimeFeatures) {
     // Generator runtime helpers for Iterator methods, send/throw, and return-value retrieval.
     generators::emit_generator_runtime(emitter);
 
-    // Array runtime functions
-    arrays::emit_heap_alloc(emitter);
-    arrays::emit_heap_debug_fail(emitter);
-    arrays::emit_heap_debug_check_live(emitter);
-    arrays::emit_heap_debug_validate_free_list(emitter);
-    arrays::emit_heap_debug_report(emitter);
-    arrays::emit_heap_kind(emitter);
-    arrays::emit_heap_free(emitter);
-    arrays::emit_array_free_deep(emitter);
-    arrays::emit_array_clone_shallow(emitter);
-    arrays::emit_array_ensure_unique(emitter);
-    arrays::emit_array_grow(emitter);
-    arrays::emit_array_new(emitter);
-    arrays::emit_array_push_int(emitter);
-    arrays::emit_array_push_refcounted(emitter);
-    arrays::emit_array_push_str(emitter);
-    arrays::emit_array_set_int(emitter);
-    arrays::emit_array_set_mixed(emitter);
-    arrays::emit_array_set_mixed_key(emitter);
-    arrays::emit_array_get_mixed_key(emitter);
-    arrays::emit_array_set_refcounted(emitter);
-    arrays::emit_array_set_str(emitter);
-    arrays::emit_array_union(emitter);
-    arrays::emit_array_hash_union(emitter);
-    arrays::emit_hash_array_union(emitter);
-    arrays::emit_random_u32(emitter);
-    arrays::emit_random_uniform(emitter);
-    arrays::emit_sort_int(emitter, false);
-    arrays::emit_sort_int(emitter, true);
-    arrays::emit_sort_str(emitter, false);
-    arrays::emit_sort_str(emitter, true);
-    arrays::emit_hash_fnv1a(emitter);
-    arrays::emit_hash_key_hash(emitter);
-    arrays::emit_hash_key_eq(emitter);
-    arrays::emit_hash_normalize_key(emitter);
-    arrays::emit_hash_clone_shallow(emitter);
-    arrays::emit_hash_ensure_unique(emitter);
-    arrays::emit_hash_new(emitter);
-    arrays::emit_hash_grow(emitter);
-    arrays::emit_hash_may_have_cyclic_values(emitter);
-    arrays::emit_hash_set(emitter);
-    arrays::emit_hash_unset(emitter);
-    arrays::emit_hash_append(emitter);
-    arrays::emit_hash_insert_owned(emitter);
-    arrays::emit_hash_get(emitter);
-    arrays::emit_hash_iter(emitter);
-    arrays::emit_hash_union(emitter);
-    arrays::emit_hash_spread(emitter);
-    arrays::emit_hash_to_mixed(emitter);
-    arrays::emit_hash_sum_mixed(emitter);
-    arrays::emit_hash_count(emitter);
-    arrays::emit_hash_free_deep(emitter);
-    arrays::emit_array_key_exists(emitter);
-    arrays::emit_undefined_array_key_warning(emitter);
-    arrays::emit_array_search(emitter);
-    arrays::emit_in_array_mixed_int(emitter);
-    arrays::emit_array_reverse(emitter);
-    arrays::emit_array_reverse_refcounted(emitter);
-    arrays::emit_array_sum(emitter);
-    arrays::emit_array_sum_mixed(emitter);
-    arrays::emit_array_product(emitter);
-    arrays::emit_array_shift(emitter);
-    arrays::emit_array_unshift(emitter);
-    arrays::emit_array_merge(emitter);
-    arrays::emit_array_merge_refcounted(emitter);
-    arrays::emit_array_slice(emitter);
-    arrays::emit_array_slice_refcounted(emitter);
-    arrays::emit_range(emitter);
-    arrays::emit_shuffle(emitter);
-    arrays::emit_array_unique(emitter);
-    arrays::emit_array_unique_refcounted(emitter);
-    arrays::emit_array_rand(emitter);
-    arrays::emit_array_fill(emitter);
-    arrays::emit_array_fill_assoc(emitter);
-    arrays::emit_array_fill_refcounted(emitter);
-    arrays::emit_array_fill_str(emitter);
-    arrays::emit_array_pad(emitter);
-    arrays::emit_array_pad_refcounted(emitter);
-    arrays::emit_array_diff(emitter);
-    arrays::emit_array_diff_refcounted(emitter);
-    arrays::emit_array_is_list(emitter);
-    arrays::emit_array_edge_key(emitter);
-    arrays::emit_array_intersect(emitter);
-    arrays::emit_array_intersect_refcounted(emitter);
-    arrays::emit_array_flip(emitter);
-    arrays::emit_array_flip_string(emitter);
-    arrays::emit_hash_flip(emitter);
-    arrays::emit_hash_map(emitter);
-    arrays::emit_array_combine(emitter);
-    arrays::emit_array_combine_refcounted(emitter);
-    arrays::emit_array_fill_keys(emitter);
-    arrays::emit_array_fill_keys_refcounted(emitter);
-    arrays::emit_array_chunk(emitter);
-    arrays::emit_array_chunk_refcounted(emitter);
-    arrays::emit_array_column(emitter);
-    arrays::emit_array_column_mixed(emitter);
-    arrays::emit_array_column_ref(emitter);
-    arrays::emit_array_column_str(emitter);
-    arrays::emit_array_splice(emitter);
-    arrays::emit_array_splice_refcounted(emitter);
-    arrays::emit_array_diff_key(emitter);
-    arrays::emit_array_intersect_key(emitter);
-    arrays::emit_array_to_hash(emitter);
-    arrays::emit_array_replace(emitter);
-    arrays::emit_array_replace_recursive(emitter);
-    arrays::emit_assoc_diff_intersect(emitter);
-    arrays::emit_amr_box_value(emitter);
-    arrays::emit_array_merge_recursive(emitter);
-    arrays::emit_array_multisort(emitter);
-    arrays::emit_asort(emitter);
-    arrays::emit_ksort(emitter);
-    arrays::emit_natsort(emitter);
-    arrays::emit_array_map(emitter);
-    arrays::emit_array_map_mixed(emitter);
-    arrays::emit_array_map_str(emitter);
-    arrays::emit_array_map_str_owned(emitter);
-    arrays::emit_array_filter(emitter);
-    arrays::emit_array_filter_refcounted(emitter);
-    arrays::emit_array_find_any_all(emitter);
-    arrays::emit_array_reduce(emitter);
-    arrays::emit_array_walk(emitter);
-    arrays::emit_array_walk_recursive(emitter);
-    arrays::emit_array_udiff_uintersect(emitter);
-    arrays::emit_usort(emitter);
-    arrays::emit_array_to_mixed(emitter);
-    arrays::emit_array_merge_into(emitter);
-    arrays::emit_array_merge_into_refcounted(emitter);
-    arrays::emit_decref_any(emitter);
-    arrays::emit_decref_mixed(emitter);
-    arrays::emit_gc_note_child_ref(emitter);
-    arrays::emit_gc_mark_reachable(emitter);
-    arrays::emit_gc_collect_cycles(emitter);
-    arrays::emit_mixed_from_value(emitter);
-    arrays::emit_mixed_abs(emitter);
-    arrays::emit_mixed_instanceof(emitter);
-    arrays::emit_iterable_unsupported_kind(emitter);
-    arrays::emit_foreach_non_iterable_warning(emitter);
-    arrays::emit_nan_bool_coercion_warning(emitter);
-    arrays::emit_iterable_write_stdout(emitter);
-    arrays::emit_mixed_cast_bool(emitter);
-    arrays::emit_mixed_cast_float(emitter);
-    arrays::emit_mixed_cast_int(emitter);
-    arrays::emit_mixed_cast_string(emitter);
-    arrays::emit_mixed_count(emitter);
-    arrays::emit_mixed_free_deep(emitter);
-    arrays::emit_mixed_is_empty(emitter);
-    arrays::emit_mixed_numeric_binops(emitter);
-    arrays::emit_int_checked_binops(emitter);
-    arrays::emit_mixed_strict_eq(emitter);
-    arrays::emit_mixed_unbox(emitter);
-    arrays::emit_mixed_write_stdout(emitter);
-    arrays::emit_object_free_deep(emitter);
-    arrays::emit_refcount(emitter);
-    if features.eval_bridge {
-        eval_bridge::emit_eval_bridge_runtime(emitter);
-    } else if features.eval_scope {
-        // Scope-only programs run compiled eval fragments natively: they need
-        // the self-contained value wrappers plus the native scope helpers
-        // (the magician staticlib supplies the scope symbols only in the full
-        // bridge configuration).
-        eval_bridge::emit_eval_bridge_runtime(emitter);
-        eval_scope::emit_eval_scope_runtime(emitter);
-    }
-
-    // SPL runtime-managed containers
-    spl::emit_doubly_linked_list_runtime(emitter);
-    spl::emit_fixed_array_runtime(emitter);
-
-    // PHP resource-id registry (its own numbering space, unrelated to object handles)
-    resource_ids::emit_resource_ids(emitter);
-
-    // Object runtime functions
-    objects::emit_object_handles(emitter);
-    objects::emit_stdclass_new(emitter);
-    objects::emit_stdclass_from_hash(emitter);
-    objects::emit_stdclass_get(emitter);
-    objects::emit_stdclass_set(emitter);
-    objects::emit_mixed_property_get(emitter);
-    objects::emit_mixed_property_set(emitter);
-    objects::emit_mixed_cell_autovivify_array(emitter);
-    objects::emit_mixed_array_get(emitter);
-    objects::emit_mixed_array_set(emitter);
-    objects::emit_mixed_array_append(emitter);
-    objects::emit_mixed_array_fetch_for_write(emitter);
-    objects::emit_new_by_name(emitter);
-    objects::emit_call_object_destructor(emitter);
-    objects::emit_json_encode_stdclass(emitter);
-
-    // Buffer runtime functions
-    buffers::emit_buffer_new(emitter);
-    buffers::emit_buffer_len(emitter);
-    buffers::emit_buffer_bounds_fail(emitter);
-    buffers::emit_buffer_use_after_free(emitter);
-
-    // I/O runtime functions
-    // The terminal-stdout indirection every echo/print travels through. Always
-    // emitted (every program can echo); its body differs for `--web` builds.
-    io::emit_stdout_write(emitter, features.web);
-    // Backs file_get_contents('php://input'); reads the request body under --web,
-    // returns false (null) otherwise. Always emitted so the EIR call resolves.
-    io::emit_php_input(emitter, features.web);
-    // Back http_response_code()/header(); call the bridge setters under --web,
-    // no-ops otherwise. Always emitted so the EIR calls resolve.
-    io::emit_http_response_code(emitter, features.web);
-    io::emit_header(emitter, features.web);
-    io::emit_cstr(emitter);
-    io::emit_disk_space(emitter);
-    io::emit_fopen(emitter);
-    io::emit_fgets(emitter);
-    io::emit_feof(emitter);
-    io::emit_stream_isatty(emitter);
-    io::emit_stream_select(emitter);
-    io::emit_stream_set_blocking(emitter);
-    io::emit_stream_set_timeout(emitter);
-    io::emit_stream_get_contents(emitter);
-    io::emit_stream_get_line(emitter);
-    io::emit_addr_is_udp(emitter);
-    io::emit_resolve_host(emitter);
-    io::emit_resolve_host_v6(emitter);
-    io::emit_inet6_pton(emitter);
-    io::emit_stream_socket_client_v6(emitter);
-    io::emit_stream_socket_server_v6(emitter);
-    io::emit_build_sockaddr_in6(emitter);
-    io::emit_opendir_glob(emitter);
-    io::emit_inet_addr_parse(emitter);
-    io::emit_format_sockaddr_in(emitter);
-    io::emit_format_sockaddr_in6(emitter);
-    io::emit_format_sockaddr_unix(emitter);
-    io::emit_data_stream(emitter);
-    io::emit_apply_stream_filter(emitter);
-    io::emit_ftp(emitter);
-    io::emit_http(emitter);
-    io::emit_https(emitter);
-    io::emit_fsockopen(emitter);
-    io::emit_stream_wrapper_register(emitter);
-    io::emit_stream_wrapper_unregister(emitter);
-    io::emit_stream_socket_server(emitter);
-    io::emit_stream_socket_client(emitter);
-    io::emit_unix_socket_server(emitter);
-    io::emit_unix_socket_client(emitter);
-    io::emit_stream_socket_accept(emitter);
-    io::emit_stream_socket_shutdown(emitter);
-    io::emit_stream_socket_sendto(emitter);
-    io::emit_stream_socket_recvfrom(emitter);
-    io::emit_stream_socket_get_name(emitter);
-    io::emit_stream_socket_pair(emitter);
-    io::emit_popen(emitter);
-    io::emit_pclose(emitter);
-    io::emit_opendir(emitter);
-    io::emit_readdir(emitter);
-    io::emit_closedir(emitter);
-    io::emit_rewinddir(emitter);
-    io::emit_stream_get_meta_data(emitter);
-    io::emit_gethostname(emitter);
-    io::emit_gethostbyname(emitter);
-    io::emit_gethostbyaddr(emitter);
-    io::emit_protoent_load(emitter);
-    io::emit_getprotobyname(emitter);
-    io::emit_getprotobynumber(emitter);
-    io::emit_servent_load(emitter);
-    io::emit_getservbyname(emitter);
-    io::emit_getservbyport(emitter);
-    io::emit_stream_copy_to_stream(emitter);
-    io::emit_stream_context_set_option_4(emitter);
-    io::emit_get_string_context_option(emitter);
-    io::emit_get_int_context_option(emitter);
-    io::emit_apply_socket_client_opts(emitter);
-    io::emit_apply_socket_server_opts(emitter);
-    io::emit_socket_backlog(emitter);
-    io::emit_apply_socket_bindto(emitter);
-    io::emit_get_ssl_peer_name(emitter);
-    io::emit_http_build_request(emitter);
-    io::emit_fread(emitter);
-    io::emit_fwrite(emitter);
-    io::emit_user_wrapper_fclose(emitter);
-    io::emit_user_wrapper_fread(emitter);
-    io::emit_user_wrapper_fwrite(emitter);
-    io::emit_user_wrapper_feof(emitter);
-    io::emit_user_wrapper_flock(emitter);
-    io::emit_user_wrapper_fseek(emitter);
-    io::emit_user_wrapper_ftell(emitter);
-    io::emit_user_wrapper_ftruncate(emitter);
-    io::emit_user_wrapper_fflush(emitter);
-    io::emit_box_wrapper_stat_result(emitter);
-    io::emit_user_wrapper_fstat(emitter);
-    io::emit_user_wrapper_url_stat(emitter);
-    io::emit_user_wrapper_url_stat_field(emitter);
-    io::emit_path_is_wrapper(emitter);
-    io::emit_readfile_wrapper(emitter);
-    io::emit_user_wrapper_path_op(emitter);
-    io::emit_user_wrapper_rename(emitter);
-    io::emit_user_wrapper_set_option(emitter);
-    io::emit_user_wrapper_opendir(emitter);
-    io::emit_user_wrapper_dir_readdir(emitter);
-    io::emit_user_wrapper_dir_closedir(emitter);
-    io::emit_user_wrapper_dir_rewinddir(emitter);
-    io::emit_touch_meta_array(emitter);
-    io::emit_stash_connect_host(emitter);
-    io::emit_fire_notification(emitter);
-    io::emit_user_wrapper_stream_cast(emitter);
-    io::emit_stream_filter_register(emitter);
-    io::emit_resolve_user_filter_id(emitter);
-    io::emit_stream_filter_attach_user(emitter);
-    io::emit_apply_user_stream_filter(emitter);
-    io::emit_user_filter_brigade_invoke(emitter);
-    io::emit_user_filter_release_fd(emitter);
-    io::emit_var_dump_array_int(emitter);
-    io::emit_var_dump_array_str(emitter);
-    io::emit_var_dump_array_bool(emitter);
-    io::emit_var_dump_array_float(emitter);
-    io::emit_var_dump_indexed(emitter);
-    io::emit_var_dump_value(emitter);
-    io::emit_var_dump_open_container(emitter);
-    io::emit_var_dump_close_container(emitter);
-    io::emit_var_dump_open_object(emitter);
-    io::emit_var_dump_object(emitter);
-    io::emit_var_dump_emit_object_key(emitter);
-    io::emit_var_dump_emit_uninit_line(emitter);
-    io::emit_var_dump_emit_recursion_line(emitter);
-    io::emit_vd_obj_desc(emitter);
-    io::emit_vd_obj_count(emitter);
-    io::emit_vd_seen_find(emitter);
-    io::emit_vd_seen_push(emitter);
-    io::emit_vd_seen_pop(emitter);
-    io::emit_var_dump_pad(emitter);
-    io::emit_var_dump_indent_step(emitter);
-    io::emit_var_dump_emit_indexed_key(emitter);
-    io::emit_var_dump_emit_string_key(emitter);
-    io::emit_var_dump_hash(emitter);
-    io::emit_var_dump_emit_int_line(emitter);
-    io::emit_var_dump_emit_string_line(emitter);
-    io::emit_var_dump_emit_bool_line(emitter);
-    io::emit_var_dump_emit_float_line(emitter);
-    io::emit_var_dump_emit_null_line(emitter);
-    io::emit_print_r_spaces(emitter);
-    io::emit_print_r_open(emitter);
-    io::emit_print_r_close(emitter);
-    io::emit_print_r_int_key(emitter);
-    io::emit_print_r_str_key(emitter);
-    io::emit_print_r_value(emitter);
-    io::emit_print_r_indexed(emitter);
-    io::emit_print_r_hash(emitter);
-    io::emit_pr_append(emitter);
-    io::emit_pr_write(emitter);
-    io::emit_pr_finish(emitter);
-    // Output-buffering (ob_*) stack helpers. Always emitted: __rt_stdout_write,
-    // __rt_pr_write, and the process-exit paths reference them unconditionally.
-    io::emit_var_dump_write(emitter);
-    io::emit_ob_start(emitter);
-    io::emit_ob_append(emitter);
-    io::emit_ob_contents(emitter);
-    io::emit_ob_queries(emitter);
-    io::emit_ob_process_and_write(emitter);
-    io::emit_ob_pop_free(emitter);
-    io::emit_ob_gated_ops(emitter);
-    io::emit_ob_get_pop_ops(emitter);
-    io::emit_ob_flush_all(emitter);
-    io::emit_ob_apply_handler(emitter);
-    io::emit_ob_result_to_bytes(emitter);
-    io::emit_ob_invoke_descriptor(emitter);
-    io::emit_ob_eval_trampoline(emitter);
-    io::emit_ob_notice_named(emitter);
-    io::emit_ob_status_entry(emitter);
-    io::emit_ob_get_status(emitter);
-    io::emit_ob_list_handlers(emitter);
-    io::emit_file_get_contents(emitter);
-    io::emit_file_put_contents(emitter);
-    io::emit_file(emitter);
-    io::emit_stat(emitter);
-    io::emit_stat_ext(emitter);
-    io::emit_stat_array(emitter);
-    io::emit_fs(emitter);
-    io::emit_getcwd(emitter);
-    io::emit_scandir(emitter);
-    io::emit_glob(emitter);
-    io::emit_tempnam(emitter);
-    io::emit_fgetcsv(emitter);
-    io::emit_fd_write(emitter);
-    io::emit_phar_write(emitter);
-    io::emit_phar_read(emitter);
-    io::emit_file_get_contents_url(emitter);
-    io::emit_fputcsv(emitter);
-    io::emit_basename(emitter);
-    io::emit_dirname(emitter);
-    io::emit_dirname_levels(emitter);
-    io::emit_fnmatch(emitter);
-    io::emit_realpath(emitter);
-    io::emit_pathinfo_str(emitter);
-    io::emit_pathinfo_array(emitter);
-    io::emit_principal_lookup(emitter);
-    io::emit_modify(emitter);
-    io::emit_streams_ext(emitter);
-    io::emit_symlink(emitter);
-
-    // Pointer runtime functions
-    pointers::emit_ptoa(emitter);
-    pointers::emit_ptr_check_nonnull(emitter);
-    pointers::emit_str_to_cstr(emitter);
-    pointers::emit_cstr_to_str(emitter);
-    pointers::emit_ptr_read_string(emitter);
-    pointers::emit_ptr_write_string(emitter);
-
-    // zval pack/unpack bridge runtime functions
-    zval::emit_zval_string_new(emitter);
-    zval::emit_zval_djbx33a(emitter);
-    zval::emit_zval_pack(emitter);
-    zval::emit_zval_pack_array_packed(emitter);
-    zval::emit_zval_pack_array_hash(emitter);
-    zval::emit_zval_unpack(emitter);
-    zval::emit_zval_unpack_array(emitter);
-    zval::emit_zval_type(emitter);
-    zval::emit_zval_free_array(emitter);
-    zval::emit_zval_free(emitter);
-
-    // Fiber runtime functions (cooperative coroutines)
-    fibers::emit_fiber_alloc_stack(emitter);
-    fibers::emit_fiber_free_stack(emitter);
-    fibers::emit_fiber_switch(emitter);
-    fibers::emit_fiber_entry(emitter);
-    fibers::emit_fiber_construct(emitter);
-    fibers::emit_fiber_throw_state_error(emitter);
-    fibers::emit_fiber_start(emitter);
-    fibers::emit_fiber_resume(emitter);
-    fibers::emit_fiber_suspend(emitter);
-    fibers::emit_fiber_throw(emitter);
-    fibers::emit_fiber_get_current(emitter);
-    fibers::emit_fiber_get_return(emitter);
-    fibers::emit_fiber_state_getter(emitter);
+    managed::emit_managed_runtime(emitter, features);
+    platform::emit_platform_runtime(emitter, features);
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::codegen_support::platform::{Arch, Platform, Target};
-
-    /// Verifies that AArch64 runtime emits fiber routines.
-    #[test]
-    fn test_aarch64_runtime_emits_fiber_routines() {
-        let mut emitter = Emitter::new(Target::new(Platform::MacOS, Arch::AArch64));
-        emit_runtime(&mut emitter, RuntimeFeatures::all());
-        let asm = emitter.output();
-
-        for sym in [
-            "__rt_fiber_alloc_stack",
-            "__rt_fiber_free_stack",
-            "__rt_fiber_switch",
-            "__rt_fiber_entry",
-            "__rt_fiber_construct",
-            "__rt_fiber_start",
-            "__rt_fiber_resume",
-            "__rt_fiber_suspend",
-            "__rt_fiber_throw",
-            "__rt_fiber_get_current",
-            "__rt_fiber_get_return",
-            "__rt_fiber_state_eq",
-        ] {
-            assert!(
-                asm.contains(&format!(".globl {}\n", sym)),
-                "fiber runtime missing global symbol {}",
-                sym
-            );
-        }
-    }
-
-    /// Verifies optional regex helpers are omitted when the program does not reference them.
-    #[test]
-    fn test_runtime_can_omit_regex_helpers() {
-        let mut emitter = Emitter::new(Target::new(Platform::MacOS, Arch::AArch64));
-        emit_runtime(&mut emitter, RuntimeFeatures::none());
-        let asm = emitter.output();
-
-        assert!(!asm.contains("__rt_preg_match:"));
-        assert!(!asm.contains("__rt_preg_replace:"));
-        assert!(!asm.contains("__rt_preg_split:"));
-    }
-
-    /// Verifies the iconv-backed `mb_strlen()` helper is emitted only for programs that use it.
-    #[test]
-    fn test_runtime_can_gate_mb_strlen_helper() {
-        let target = Target::new(Platform::MacOS, Arch::AArch64);
-        let mut omitted = Emitter::new(target);
-        emit_runtime(&mut omitted, RuntimeFeatures::none());
-        assert!(!omitted.output().contains("__rt_mb_strlen:"));
-
-        let mut included = Emitter::new(target);
-        emit_runtime(
-            &mut included,
-            RuntimeFeatures {
-                mb_strlen: true,
-                ..RuntimeFeatures::none()
-            },
-        );
-        assert!(included.output().contains("__rt_mb_strlen:"));
-    }
-
-    /// Verifies that Linux x86_64 uses the shared runtime surface.
-    #[test]
-    fn test_linux_x86_64_runtime_uses_shared_surface() {
-        let mut emitter = Emitter::new(Target::new(Platform::Linux, Arch::X86_64));
-        emit_runtime(&mut emitter, RuntimeFeatures::all());
-        let asm = emitter.output();
-
-        for sym in [
-            "__rt_hash_count",
-            "__rt_gc_note_child_ref",
-            "__rt_incref",
-            "__rt_decref_array",
-            "__rt_json_encode_assoc",
-            "__rt_preg_match",
-            "__rt_fiber_alloc_stack",
-        ] {
-            assert!(
-                asm.contains(&format!(".globl {}\n", sym)),
-                "linux x86_64 shared runtime missing global symbol {}",
-                sym
-            );
-        }
-    }
-
-    /// Verifies the full macOS AArch64 runtime still assembles once per-symbol
-    /// dead stripping is enabled. The real codegen path renames internal labels
-    /// to `L`-locals and appends a `.subsections_via_symbols` footer; under that
-    /// mode the Mach-O assembler rejects any conditional branch whose target is
-    /// another atom (another helper) or a non-local label. Assembling the
-    /// all-features runtime catches every such cross-helper conditional branch
-    /// at build time rather than letting it slip into a miscompiled binary.
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_macos_dead_strip_runtime_assembles() {
-        // Use the real runtime generation path (pic = false → macOS executable),
-        // so the assembly is exactly what is linked, including label localization.
-        let asm = crate::codegen_support::generate_runtime_with_features_pic(
-            8 * 1024 * 1024,
-            Target::new(Platform::MacOS, Arch::AArch64),
-            RuntimeFeatures::all(),
-            false,
-        );
-
-        let dir = std::env::temp_dir().join(format!(
-            "elephc_deadstrip_asm_{}",
-            std::process::id()
-        ));
-        std::fs::create_dir_all(&dir).expect("create temp dir");
-        let asm_path = dir.join("runtime.s");
-        let obj_path = dir.join("runtime.o");
-        std::fs::write(&asm_path, &asm).expect("write asm");
-
-        let output = std::process::Command::new("as")
-            .args(["-arch", "arm64", "-o"])
-            .arg(&obj_path)
-            .arg(&asm_path)
-            .output()
-            .expect("run as");
-        let _ = std::fs::remove_dir_all(&dir);
-
-        assert!(
-            output.status.success(),
-            "macOS dead-strip runtime failed to assemble:\n{}",
-            String::from_utf8_lossy(&output.stderr)
-        );
-    }
-
-    /// Guards the atom invariant the assemble-only test cannot see: under macOS
-    /// `-dead_strip` an internal helper label is renamed to an `L`-local, which
-    /// is not a symbol, so a reference to it from *another* atom (helper) is not
-    /// a relocation the linker can follow. The target atom is then stripped even
-    /// though a live atom still branches into it, miscompiling silently — this is
-    /// the bug that made `foreach` over an associative array crash in
-    /// `__rt_mixed_unbox`. A cross-helper helper must instead use `label_shared`
-    /// (`.alt_entry`) so it stays a real symbol inside its atom.
-    ///
-    /// This parses the real dead-strip runtime and asserts every `L__rt_*`
-    /// reference resolves within its defining atom. `.alt_entry` labels stay bare
-    /// (not `L`-localized) so they are correctly excluded; numeric local labels
-    /// never start an atom and are ignored.
-    #[test]
-    #[cfg(target_os = "macos")]
-    fn test_macos_dead_strip_no_cross_atom_internal_refs() {
-        let asm = crate::codegen_support::generate_runtime_with_features_pic(
-            8 * 1024 * 1024,
-            Target::new(Platform::MacOS, Arch::AArch64),
-            RuntimeFeatures::all(),
-            false,
-        );
-
-        // A token is an internal helper label iff it is an `L`-localized `__rt_*`
-        // name (what `label()` produces under dead stripping). `.alt_entry`
-        // helpers stay bare `__rt_*`, so they never match here.
-        /// Returns whether an assembly token names a dead-strip-local runtime helper.
-        fn is_internal(tok: &str) -> bool {
-            tok.starts_with("L__rt_")
-        }
-        // True when `s` is a bare label definition body (no whitespace, label
-        // characters only, not purely numeric → not an assembler-local `N:`).
-        /// Returns whether a token can be a non-numeric assembly label definition.
-        fn is_label_name(s: &str) -> bool {
-            !s.is_empty()
-                && !s.bytes().all(|b| b.is_ascii_digit())
-                && s.bytes()
-                    .all(|b| b.is_ascii_alphanumeric() || matches!(b, b'_' | b'$' | b'.'))
-        }
-
-        let mut current_atom: &str = "<root>";
-        let mut prev_alt_entry: Option<&str> = None;
-        let mut owner: std::collections::HashMap<&str, &str> = std::collections::HashMap::new();
-        let mut refs: Vec<(&str, &str)> = Vec::new();
-
-        for raw in asm.lines() {
-            let line = raw.trim();
-            if line.is_empty() || line.starts_with("//") {
-                continue;
-            }
-            if let Some(rest) = line.strip_prefix(".alt_entry ") {
-                prev_alt_entry = Some(rest.trim());
-                continue;
-            }
-            // Label definition: a single `name:` token on the line.
-            if let Some(name) = line.strip_suffix(':') {
-                if is_label_name(name) {
-                    if is_internal(name) {
-                        owner.insert(name, current_atom);
-                    } else if prev_alt_entry != Some(name) {
-                        // A real global symbol starts a new atom; an `.alt_entry`
-                        // label stays inside the current atom (not a boundary).
-                        current_atom = name;
-                    }
-                }
-                prev_alt_entry = None;
-                continue;
-            }
-            prev_alt_entry = None;
-            // Reference scan: collect `L__rt_*` tokens used as operands.
-            for tok in line.split(|c: char| !(c.is_ascii_alphanumeric() || matches!(c, '_' | '$' | '.'))) {
-                if is_internal(tok) {
-                    refs.push((current_atom, tok));
-                }
-            }
-        }
-
-        let mut violations: Vec<String> = refs
-            .iter()
-            .filter_map(|(atom, tok)| match owner.get(tok) {
-                Some(def_atom) if def_atom != atom => Some(format!(
-                    "{tok} defined in {def_atom} but referenced from {atom}"
-                )),
-                _ => None,
-            })
-            .collect();
-        violations.sort();
-        violations.dedup();
-        assert!(
-            violations.is_empty(),
-            "cross-atom references to internal `__rt_*` labels would be stripped \
-             under -dead_strip (use label_shared/.alt_entry for cross-helper \
-             targets):\n{}",
-            violations.join("\n")
-        );
-    }
-}
+mod tests;
