+++ src/codegen_support/runtime/emitters/managed.rs
@@ -1,0 +1,209 @@
+//! Purpose:
+//! Emits runtime helpers for managed PHP values, containers, resources, objects, and buffers.
+//!
+//! Called from:
+//! - `super::emit_runtime()` after the core runtime helpers.
+//!
+//! Key details:
+//! - Keeps heap, GC, eval-scope, SPL, object, and buffer helpers in dependency order.
+
+use super::super::{arrays, buffers, eval_bridge, eval_scope, objects, resource_ids, spl};
+use crate::codegen_support::emit::Emitter;
+use crate::codegen_support::RuntimeFeatures;
+
+/// Emits helpers for managed values and their container or object abstractions.
+pub(super) fn emit_managed_runtime(emitter: &mut Emitter, features: RuntimeFeatures) {
+    // Array runtime functions
+    arrays::emit_heap_alloc(emitter);
+    arrays::emit_heap_debug_fail(emitter);
+    arrays::emit_heap_debug_check_live(emitter);
+    arrays::emit_heap_debug_validate_free_list(emitter);
+    arrays::emit_heap_debug_report(emitter);
+    arrays::emit_heap_kind(emitter);
+    arrays::emit_heap_free(emitter);
+    arrays::emit_array_free_deep(emitter);
+    arrays::emit_array_clone_shallow(emitter);
+    arrays::emit_array_ensure_unique(emitter);
+    arrays::emit_array_grow(emitter);
+    arrays::emit_array_new(emitter);
+    arrays::emit_array_push_int(emitter);
+    arrays::emit_array_push_refcounted(emitter);
+    arrays::emit_array_push_str(emitter);
+    arrays::emit_array_set_int(emitter);
+    arrays::emit_array_set_mixed(emitter);
+    arrays::emit_array_set_mixed_key(emitter);
+    arrays::emit_array_get_mixed_key(emitter);
+    arrays::emit_array_set_refcounted(emitter);
+    arrays::emit_array_set_str(emitter);
+    arrays::emit_array_union(emitter);
+    arrays::emit_array_hash_union(emitter);
+    arrays::emit_hash_array_union(emitter);
+    arrays::emit_random_u32(emitter);
+    arrays::emit_random_uniform(emitter);
+    arrays::emit_sort_int(emitter, false);
+    arrays::emit_sort_int(emitter, true);
+    arrays::emit_sort_str(emitter, false);
+    arrays::emit_sort_str(emitter, true);
+    arrays::emit_hash_fnv1a(emitter);
+    arrays::emit_hash_key_hash(emitter);
+    arrays::emit_hash_key_eq(emitter);
+    arrays::emit_hash_normalize_key(emitter);
+    arrays::emit_hash_clone_shallow(emitter);
+    arrays::emit_hash_ensure_unique(emitter);
+    arrays::emit_hash_new(emitter);
+    arrays::emit_hash_grow(emitter);
+    arrays::emit_hash_may_have_cyclic_values(emitter);
+    arrays::emit_hash_set(emitter);
+    arrays::emit_hash_unset(emitter);
+    arrays::emit_hash_append(emitter);
+    arrays::emit_hash_insert_owned(emitter);
+    arrays::emit_hash_get(emitter);
+    arrays::emit_hash_iter(emitter);
+    arrays::emit_hash_union(emitter);
+    arrays::emit_hash_spread(emitter);
+    arrays::emit_hash_to_mixed(emitter);
+    arrays::emit_hash_sum_mixed(emitter);
+    arrays::emit_hash_count(emitter);
+    arrays::emit_hash_free_deep(emitter);
+    arrays::emit_array_key_exists(emitter);
+    arrays::emit_undefined_array_key_warning(emitter);
+    arrays::emit_array_search(emitter);
+    arrays::emit_in_array_mixed_int(emitter);
+    arrays::emit_array_reverse(emitter);
+    arrays::emit_array_reverse_refcounted(emitter);
+    arrays::emit_array_sum(emitter);
+    arrays::emit_array_sum_mixed(emitter);
+    arrays::emit_array_product(emitter);
+    arrays::emit_array_shift(emitter);
+    arrays::emit_array_unshift(emitter);
+    arrays::emit_array_merge(emitter);
+    arrays::emit_array_merge_refcounted(emitter);
+    arrays::emit_array_slice(emitter);
+    arrays::emit_array_slice_refcounted(emitter);
+    arrays::emit_range(emitter);
+    arrays::emit_shuffle(emitter);
+    arrays::emit_array_unique(emitter);
+    arrays::emit_array_unique_refcounted(emitter);
+    arrays::emit_array_rand(emitter);
+    arrays::emit_array_fill(emitter);
+    arrays::emit_array_fill_assoc(emitter);
+    arrays::emit_array_fill_refcounted(emitter);
+    arrays::emit_array_fill_str(emitter);
+    arrays::emit_array_pad(emitter);
+    arrays::emit_array_pad_refcounted(emitter);
+    arrays::emit_array_diff(emitter);
+    arrays::emit_array_diff_refcounted(emitter);
+    arrays::emit_array_is_list(emitter);
+    arrays::emit_array_edge_key(emitter);
+    arrays::emit_array_intersect(emitter);
+    arrays::emit_array_intersect_refcounted(emitter);
+    arrays::emit_array_flip(emitter);
+    arrays::emit_array_flip_string(emitter);
+    arrays::emit_hash_flip(emitter);
+    arrays::emit_hash_map(emitter);
+    arrays::emit_array_combine(emitter);
+    arrays::emit_array_combine_refcounted(emitter);
+    arrays::emit_array_fill_keys(emitter);
+    arrays::emit_array_fill_keys_refcounted(emitter);
+    arrays::emit_array_chunk(emitter);
+    arrays::emit_array_chunk_refcounted(emitter);
+    arrays::emit_array_column(emitter);
+    arrays::emit_array_column_mixed(emitter);
+    arrays::emit_array_column_ref(emitter);
+    arrays::emit_array_column_str(emitter);
+    arrays::emit_array_splice(emitter);
+    arrays::emit_array_splice_refcounted(emitter);
+    arrays::emit_array_diff_key(emitter);
+    arrays::emit_array_intersect_key(emitter);
+    arrays::emit_array_to_hash(emitter);
+    arrays::emit_array_replace(emitter);
+    arrays::emit_array_replace_recursive(emitter);
+    arrays::emit_assoc_diff_intersect(emitter);
+    arrays::emit_amr_box_value(emitter);
+    arrays::emit_array_merge_recursive(emitter);
+    arrays::emit_array_multisort(emitter);
+    arrays::emit_asort(emitter);
+    arrays::emit_ksort(emitter);
+    arrays::emit_natsort(emitter);
+    arrays::emit_array_map(emitter);
+    arrays::emit_array_map_mixed(emitter);
+    arrays::emit_array_map_str(emitter);
+    arrays::emit_array_map_str_owned(emitter);
+    arrays::emit_array_filter(emitter);
+    arrays::emit_array_filter_refcounted(emitter);
+    arrays::emit_array_find_any_all(emitter);
+    arrays::emit_array_reduce(emitter);
+    arrays::emit_array_walk(emitter);
+    arrays::emit_array_walk_recursive(emitter);
+    arrays::emit_array_udiff_uintersect(emitter);
+    arrays::emit_usort(emitter);
+    arrays::emit_array_to_mixed(emitter);
+    arrays::emit_array_merge_into(emitter);
+    arrays::emit_array_merge_into_refcounted(emitter);
+    arrays::emit_decref_any(emitter);
+    arrays::emit_decref_mixed(emitter);
+    arrays::emit_gc_note_child_ref(emitter);
+    arrays::emit_gc_mark_reachable(emitter);
+    arrays::emit_gc_collect_cycles(emitter);
+    arrays::emit_mixed_from_value(emitter);
+    arrays::emit_mixed_abs(emitter);
+    arrays::emit_mixed_instanceof(emitter);
+    arrays::emit_iterable_unsupported_kind(emitter);
+    arrays::emit_foreach_non_iterable_warning(emitter);
+    arrays::emit_nan_bool_coercion_warning(emitter);
+    arrays::emit_iterable_write_stdout(emitter);
+    arrays::emit_mixed_cast_bool(emitter);
+    arrays::emit_mixed_cast_float(emitter);
+    arrays::emit_mixed_cast_int(emitter);
+    arrays::emit_mixed_cast_string(emitter);
+    arrays::emit_mixed_count(emitter);
+    arrays::emit_mixed_free_deep(emitter);
+    arrays::emit_mixed_is_empty(emitter);
+    arrays::emit_mixed_numeric_binops(emitter);
+    arrays::emit_int_checked_binops(emitter);
+    arrays::emit_mixed_strict_eq(emitter);
+    arrays::emit_mixed_unbox(emitter);
+    arrays::emit_mixed_write_stdout(emitter);
+    arrays::emit_object_free_deep(emitter);
+    arrays::emit_refcount(emitter);
+    if features.eval_bridge {
+        eval_bridge::emit_eval_bridge_runtime(emitter);
+    } else if features.eval_scope {
+        // Scope-only programs run compiled eval fragments natively: they need
+        // the self-contained value wrappers plus the native scope helpers
+        // (the magician staticlib supplies the scope symbols only in the full
+        // bridge configuration).
+        eval_bridge::emit_eval_bridge_runtime(emitter);
+        eval_scope::emit_eval_scope_runtime(emitter);
+    }
+
+    // SPL runtime-managed containers
+    spl::emit_doubly_linked_list_runtime(emitter);
+    spl::emit_fixed_array_runtime(emitter);
+
+    // PHP resource-id registry (its own numbering space, unrelated to object handles)
+    resource_ids::emit_resource_ids(emitter);
+
+    // Object runtime functions
+    objects::emit_object_handles(emitter);
+    objects::emit_stdclass_new(emitter);
+    objects::emit_stdclass_from_hash(emitter);
+    objects::emit_stdclass_get(emitter);
+    objects::emit_stdclass_set(emitter);
+    objects::emit_mixed_property_get(emitter);
+    objects::emit_mixed_property_set(emitter);
+    objects::emit_mixed_cell_autovivify_array(emitter);
+    objects::emit_mixed_array_get(emitter);
+    objects::emit_mixed_array_set(emitter);
+    objects::emit_mixed_array_append(emitter);
+    objects::emit_mixed_array_fetch_for_write(emitter);
+    objects::emit_new_by_name(emitter);
+    objects::emit_call_object_destructor(emitter);
+    objects::emit_json_encode_stdclass(emitter);
+
+    // Buffer runtime functions
+    buffers::emit_buffer_new(emitter);
+    buffers::emit_buffer_len(emitter);
+    buffers::emit_buffer_bounds_fail(emitter);
+    buffers::emit_buffer_use_after_free(emitter);
+}
